Conflicts:
	app/views/pages/product/create.blade.php

+++ app/controllers/ProductController.php
@@ -1,12 +1,15 @@
 <?php
 
-class ProductController extends \BaseController {
+class ProductController extends BaseController {
 
 	/**
 	 * Display a listing of the resource.
 	 *
 	 * @return Response
 	 */
+	public function __construct() {
+		$this->beforeFilter('csrf', array('on'=>'post'));
+	}
 	public function index()
 	{
 		$products = Prod::All();	
@@ -53,8 +56,13 @@ class ProductController extends \BaseController {
 			$product->price         = Input::get('price');
 			$product->brand_id      = Input::get('brand');
 			$product->category_id   = Input::get('category');
-			$product->save();
+			$image = Input::file('product_pic');
+			$filename = date('Y-m-d-H-i-s')."-".$image->getClientOriginalName();
+			Image::make($image->getRealPath())->save(public_path().'/img/products/'.$filename);
 
+			$product->product_pic = $filename;
+			$product->save();
+			var_dump($product);
 			// redirect
 			Session::flash('message', 'Successfully created product!');
 			return Redirect::to('product');
@@ -120,6 +128,12 @@ class ProductController extends \BaseController {
 			$product->price         = Input::get('price');
 			$product->brand_id      = Input::get('brand');
 			$product->category_id   = Input::get('category');
+			File::delete(public_path().$product->image);
+
+			$image = Input::file('product_pic');
+			$filename = date('Y-m-d-H-i-s')."-".$image->getClientOriginalName();
+			Image::make($image->getRealPath())->save(public_path().'/img/products/'.$filename);
+			$product->product_pic = $filename;
 			$product->save();
 
 			// redirect
@@ -137,10 +151,16 @@ class ProductController extends \BaseController {
 	public function destroy($id)
 	{
 		$product = Prod::find($id);
-		$product->delete();
+		if ($product) {
+			File::delete(public_path().$product->image);
+			$product->delete();
+			Session::flash('message', 'Successfully deleted the product!');
+		return Redirect::to('product');
+		}
+		
 
 		// redirect
-		Session::flash('message', 'Successfully deleted the product!');
+		Session::flash('message', 'Something went wrong, please try again');
 		return Redirect::to('product');
 	}
